transaction logging and exception handling

+++ src/main/java/com/opensource/cloudnest/repository/ProfileRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProfileRepository extends JpaRepository<Profile, Long> {
+public interface ProfileRepository extends JpaRepository<Profile, Integer> {
 Optional<Profile> findByEmail(String email);
 Optional<Profile> findByUserName(String userName);
